+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.2]
 
-* TODO: Describe initial release.
+* Revamped example
+* Improved description
+
+## [0.0.1] - Initial release
+
+* Fully working PaginatedConsumer; example **included**